Financial_market_simulator/requirements.txt
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
psycopg2==2.9.3
Jinja2==3.0.3
Werkzeug==2.0.3
itsdangerous==2.0.1
click==8.0.3
